plugins/requirements.txt
GoogleAppEngineCloudStorageClient
influxdb
ijson
requests-toolbelt
